New deconvolution code
diffuser_2d_deconv_v2.m is a cleaned up 2D deconvolution code that uses
the standard linear gradient calculation, as well as function handles
for A and A'. Additionally, a wrapper called realtime_deconv.m will poll
a directory for any image file, load it into memory then deconvolve. the
result will be moved to the server and timestamped.